moved merge macro to its own module
